chore: reorg states/mod
